API: Restructure raster_scan3D to start mid-procedure faster
Instead of calculating each probe before the desired start point,
raster_scan3D scan now jump to an arbitrary measurement in the
procedure quickly.

BUG: Remove np.nan to avoid Warning

Using np.nan created a warning in Mesh.bounding_box. By replacing
nan with inf and -inf, the warning is avoided. For some reason
the warning was not visible in earlier versions of numpy. I don't
know why it appeared because fmin is supposed to ignore the fact
that nans exist.